made walkthrough so it does not respond to voice commands when in development mode
This change came about because I was dictating voice commands while in development mode and the walkthrough was responding as if I was actually saying those commands and highlighting them green and then automatically going to the next at. This was particularly annoying when I am adding as To the end of the tutorial because automatically going to the next step at the end of the editorial actually closes out the tutorial and I have to reopen it.